feat: Improve database setup/config/platform compatibility and restructure documentation

+++ migrations/20230604031458-clear_channels_table.js
@@ -1,8 +1,14 @@
 'use strict';
 
+const { tableExists } = require('./helpers');
+
 /** @type {import('sequelize-cli').Migration} */
 module.exports = {
   async up(queryInterface, Sequelize) {
+    if (!(await tableExists(queryInterface, 'channels'))) {
+      return;
+    }
+
     await queryInterface.bulkDelete('channels', null, {});
   },
